改写MCU中的矩阵单位由int -> uint16 or int16 （使用int8 uint8有问题）

+++ include/Common.hpp
@@ -36,18 +36,17 @@ inline const bool checkSpace(const string &value) {
   return true;
 }
 
-inline void
-printMatrix(const array<array<int, COMPONENT_SIZE>, COMPONENT_SIZE> matrix) {
+inline void printMatrix(
+    const array<array<uint8_t, COMPONENT_SIZE>, COMPONENT_SIZE> matrix) {
   int rows = COMPONENT_SIZE, cols = COMPONENT_SIZE;
   for (int i = 0; i < rows; ++i) {
     std::cout << "|";
     for (int j = 0; j < cols; ++j)
-      std::cout << setw(4) << matrix[i][j];
+      std::cout << setw(4) << (int)matrix[i][j];
     std::cout << "|" << std::endl;
   }
 }
 
-/* TODO YangJing 这里有点混乱，我认为应该都使用uint8_t <24-04-30 16:07:41> */
 template <typename T>
 int arrayToMatrixUseZigZag(
     const array<T, MCU_UNIT_SIZE> a,
@@ -60,7 +59,6 @@ template <typename T>
 int matrixToArrayUseZigZag(
     const array<array<T, COMPONENT_SIZE>, COMPONENT_SIZE> &matrix,
     array<T, MCU_UNIT_SIZE> &arr);
-/* TODO YangJing 这里有点混乱，我认为应该都使用uint8_t <24-04-30 16:07:41> */
 
 inline string getFileType(string &fileName) {
   return fileName.substr(fileName.find_last_of('.') + 1);
